Split tests for CSV and JSON subplot downloads into dedicated files
